empty onMount instead of double get
RollNewNames tries to get the link to the nameGen however on first paint it is not yet reachable, for some reason?

Instead of doing it twice, just dispatch the empty result so that the cards are painted, then roll new names after a pause. Prevents double calling. Less traffic = better