Remove random-Name & npm update & v2
random-name was a side loaded node.js applet which supplied randomised names. However the way this was required to run and how we got the names from it was strange.
Also depended on another developer to adjust as I didn't understand it.

npm update to bring everything to now.

Version number changed to 2.0.0 as these changes break any backwards compatibility that may have been present.