Switch to JSON
Instead of hardcoding names and variables, we switch to json data.
This is not possible everywhere, and at this time we create the array 3 times because I do not yet know how Svelte Stores work regarding shared arrays.

Once that's figured out, a single array could be shared across all these functions.