style checkbox + hide from non spelleider
The checkboxes in the figu overview are now styled - the modal checkbox style has been altered to be the same style.

There now is a variable that can be set with a string to 'spelleider' (or something else). When it is anything but spelleider, all management buttons vanish. Getting the user type is not yet a function at this time.