Prelaunch adjustments (#27)
* Planet no longer always Eos

The planet was always Eos due to declaring the value at the top. This somehow disallowed other changes. It now behaves properly again, and has reduced the error log still, as intended for this convolution.

* style checkbox + hide from non spelleider

The checkboxes in the figu overview are now styled - the modal checkbox style has been altered to be the same style.

There now is a variable that can be set with a string to 'spelleider' (or something else). When it is anything but spelleider, all management buttons vanish. Getting the user type is not yet a function at this time.

* name generator branding

The title of the website now also changes from 'figumanager' to 'Name Generator'. Still requires its own favIcon that can also be used in the header.

* asign figurant on generation

A figurant can now be assigned at creation! Whoo!

* meta data

DOES NOT WORK - ask Josh for extra field WAY WAY easier to implement

* export tags

assigned figu & plot font awesome icon tags

* edit data modal

The edit data modal has its basic logic structure set up. The data is correctly sent to the component. The component itself is lacking an interface atm.

* edit screen with data

the data is now presented into the edit screen - the save button does not yet work

* get currentYear from watchtower

instead of hardcoding the current year, watchtower now supplies the year.

* watchtower currentYear for generation too

The generation screen now also gets its current year for birthday calculation from watchtower.

* renamed let so it is clear the year is in character

* use better RegEx against hardcoding

Different year formats are now supported, instead of only 3 digit years

* added planets and data can now be changed (not yet saved)

* edit works

edit knopje werkt en doet het scherm verversen

* handle unsetting

a figu char can now be unasigned from an figu account id proper.

* null != 'null'

string vs actual null. Weird handling keep it as string

* fix refresh button

the refresh button had an unclear icon (did not load on all platforms)

* RFID card altering

The code is present, but the database is not complying yet. Maybe code bug?

* savesuccess for RFID

refresh datatable if rfid card is altered

* birthday is now saved

forgot the brithday field in the edit figurant screen save state

* different service-worker script

This one can handle the pass through of get's and puts. Eliminating a lot of non-responsiveness on first load

* rfid checks if changed

The rfid on the persona table now checks if it actually was changed before it sends it off.

It also checks if it is null and empty to prevent conflict.

* user identifycation + deprication + catch

When watchtower is offline this is now caught

users are identified when a joomla session is supplied (does not work yet fullly, needs joomla session)

removed 2 files that are no longer needed

* rescaled header

The header title is now properly scaling on smaller devices.

* idandgroups

Instead of relying on orthanc we use the idandgroups.php file to resolve the joomla id.

* service worker caches renamed

for proper caching refresh

* Hide login if it is a speler that is logged in

Login was always present even if the joomla id was detected, and only hidden for spelleiders and figuranten

* plotname

plotname datakey has been added

* worker cache + debug

removed debug line and renamed caches for showtime

* renamed cache

cache renamed for api test